make separate file for containers proxy
